Updates to bowling and api

+++ TruSport/Model/BowlingScore.cs
@@ -7,18 +7,18 @@ namespace TruSport.Model
     {
         public string ID { get; set; }
         public string BowlingFixtureID { get; set; }
-        public int? HomeTeamPoints { get; set; }
-        public int? AwayTeamPoints { get; set; }
-        public int? HomeTeamMatchPoints { get; set; }
-        public int? AwayTeamMatchPoints { get; set; }
+        public decimal? HomeTeamPoints { get; set; }
+        public decimal? AwayTeamPoints { get; set; }
+        public decimal? HomeTeamMatchPoints { get; set; }
+        public decimal? AwayTeamMatchPoints { get; set; }
 
         [Ignore]
         public BowlingFixture BowlingFixture { get; set; }
 
         [Ignore]
-        public int? HomeTeamTotalPoints { get; set; }
+        public decimal? HomeTeamTotalPoints { get; set; }
 
         [Ignore]
-        public int? AwayTeamTotalPoints { get; set; }
+        public decimal? AwayTeamTotalPoints { get; set; }
     }
 }
